Movie adding implemented

+++ Cinema/BusinessLogic/DTOs/MovieDTO.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public string Title { get; set; }
         public string Description { get; set; }
-        public ushort Year { get; set; }
+        public int Year { get; set; }
         public float Rating { get; set; }
         public string TrailerLink { get; set; }
         public string ThumbnailLink { get; set; }
